refactoring 共通のscriptを使いたいがために、共通のコンポーネント使いHTML部分を条件分岐で無理やり切り替えていたが、mixinを使うことでscriptだけを共通化させることができたのでmixinを使うように変更
