+++ src/assets/styles.scss
@@ -41,7 +41,7 @@ a:visited {
   color: #999999;
 }
 
-.music-info__music__lists__lyric, .music-info__lyrics__lists__ele__lyric {
+.music__lists__lyric, .lyrics__lists__ele__lyric {
   background-image: url(../assets/listBackgroundImage.jpg);
   background-size: cover;
   font-size: 90%;
